Convertex line buckets
